pyproject.toml
[tool.poetry]
name = "pptx-extractor"
version = "0.1.0"
description = "An app to scrape text and images from pptx files."
authors = ["mattyocode <matthewoliver@live.co.uk>"]
readme = "README.md"
homepage = "https://github.com/mattyocode/pptx-scraper.git"
repository = "https://github.com/mattyocode/pptx-scraper.git"
keywords = ["pptx", "scraper"]

[tool.poetry.dependencies]
python = "^3.7"
python-pptx = "^0.6.18"
SQLAlchemy = "^1.3.22"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
